Refactor project into src package and add entry script
